Refactor entire codebase
